requirements.txt
pandas
xlrd
valentine
translators
python-Levenshtein
recordlinkage